GA-133 Split parser for AccessTypes and AccessTypeRoles
