+++ excel-importer/src/main/java/uk/gov/hmcts/ccd/definition/store/excel/validation/AccessTypesValidator.java
@@ -1,0 +1,151 @@
+package uk.gov.hmcts.ccd.definition.store.excel.validation;
+
+import com.microsoft.applicationinsights.boot.dependencies.apachecommons.lang3.tuple.Triple;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+import uk.gov.hmcts.ccd.definition.store.domain.validation.ValidationError;
+import uk.gov.hmcts.ccd.definition.store.domain.validation.ValidationResult;
+import uk.gov.hmcts.ccd.definition.store.excel.util.mapper.ColumnName;
+import uk.gov.hmcts.ccd.definition.store.excel.util.mapper.SheetName;
+import uk.gov.hmcts.ccd.definition.store.repository.entity.AccessTypeEntity;
+import uk.gov.hmcts.ccd.definition.store.repository.entity.CaseTypeEntity;
+
+import java.util.List;
+import java.util.Map;
+
+import static java.util.stream.Collectors.groupingBy;
+
+@Component
+public class AccessTypesValidator {
+
+    public ValidationResult validate(final List<AccessTypeEntity> accessTypeEntities) {
+
+        ValidationResult validationResult = new ValidationResult();
+
+        accessTypeEntities
+            .forEach(entity -> {
+                validateRequiredFields(validationResult, entity, accessTypeEntities);
+                validateDisplay(validationResult, entity, accessTypeEntities);
+            });
+
+        return validationResult;
+    }
+
+    private void validateRequiredFields(ValidationResult validationResult, AccessTypeEntity entity,
+                                        List<AccessTypeEntity> accessTypeEntities) {
+
+        if (!StringUtils.hasLength(entity.getAccessTypeId())) {
+            validationResult.addError(new ValidationError(
+                String.format("Access Type ID should not be null or empty in column '%s' in the sheet '%s'",
+                    ColumnName.ACCESS_TYPE_ID, SheetName.ACCESS_TYPE_ROLES)) {
+            });
+        } else {
+            validateAccessTypeId(validationResult, accessTypeEntities);
+        }
+
+        if (!StringUtils.hasLength(entity.getOrganisationProfileId())) {
+            validationResult.addError(new ValidationError(
+                String.format("Organisation Profile ID should not be null or empty in column '%s' "
+                    + "in the sheet '%s'", ColumnName.ORGANISATION_PROFILE_ID, SheetName.ACCESS_TYPE_ROLES)) {
+            });
+        }
+    }
+
+    private void validateAccessTypeId(ValidationResult validationResult,
+                                      List<AccessTypeEntity> accessTypeEntities) {
+
+        Map<Triple<CaseTypeEntity, Integer, String>, List<AccessTypeEntity>> accessTypeRolesAccessTypeId =
+            accessTypeEntities
+                .stream()
+                .collect(groupingBy(p ->
+                    Triple.of(p.getCaseTypeId(),
+                        p.getCaseTypeId().getJurisdiction().getId(),
+                        p.getAccessTypeId())));
+
+        accessTypeRolesAccessTypeId.keySet()
+            .forEach(triple -> {
+                if (accessTypeRolesAccessTypeId.get(triple).size() > 1) {
+                    String errorMessage = String.format("'%s' must be unique within the Jurisdiction in the sheet '%s'",
+                        ColumnName.ACCESS_TYPE_ID, SheetName.ACCESS_TYPE_ROLES);
+
+                    if (!alreadyReportedError(validationResult, errorMessage)) {
+                        validationResult.addError(new ValidationError(errorMessage) {});
+                    }
+                }
+            });
+    }
+
+    private void validateDisplay(ValidationResult validationResult, AccessTypeEntity entity,
+                                 List<AccessTypeEntity> accessTypeEntities) {
+
+        if (entity.getDisplay()) {
+            if (!entity.getAccessMandatory() || !entity.getAccessDefault()) {
+                validationResult.addError(new ValidationError(
+                    String.format("'%s' and '%s' must be set to true for '%s' to be used in the sheet '%s'",
+                        ColumnName.ACCESS_MANDATORY, ColumnName.ACCESS_DEFAULT,
+                        ColumnName.DISPLAY, SheetName.ACCESS_TYPE_ROLES)) {
+                });
+            }
+
+            if (!StringUtils.hasLength(entity.getDescription())) {
+                validationResult.addError(new ValidationError(
+                    String.format("'%s' must be set for '%s' to be used in the sheet '%s'",
+                        ColumnName.DESCRIPTION, ColumnName.DISPLAY, SheetName.ACCESS_TYPE_ROLES)) {
+                });
+            }
+
+            if (!StringUtils.hasLength(entity.getHint())) {
+                validationResult.addError(new ValidationError(
+                    String.format("'%s' must be set for '%s' to be used in the sheet '%s'",
+                        ColumnName.HINT_TEXT, ColumnName.DISPLAY, SheetName.ACCESS_TYPE_ROLES)) {
+                });
+            }
+
+            if (entity.getDisplayOrder() == null) {
+                validationResult.addError(new ValidationError(
+                    String.format("'%s' should not be null or empty for '%s' to be used in column '%s' "
+                            + "in the sheet '%s'", ColumnName.DISPLAY_ORDER, ColumnName.DISPLAY,
+                        ColumnName.DISPLAY_ORDER, SheetName.ACCESS_TYPE_ROLES)) {
+                });
+
+            } else if (entity.getDisplayOrder() < 1) {
+                validationResult.addError(new ValidationError(
+                    String.format("'%s' must be greater than 0 in column '%s' in the sheet '%s'",
+                        ColumnName.DISPLAY_ORDER, ColumnName.DISPLAY_ORDER, SheetName.ACCESS_TYPE_ROLES)) {
+                });
+            } else {
+                validateDisplayOrder(validationResult, accessTypeEntities);
+            }
+        }
+    }
+
+    private void validateDisplayOrder(ValidationResult validationResult,
+                                      List<AccessTypeEntity> accessTypeEntities) {
+
+        Map<Triple<CaseTypeEntity, Integer, Integer>, List<AccessTypeEntity>> accessTypeRolesDisplayOrder =
+            accessTypeEntities
+                .stream()
+                .collect(groupingBy(p ->
+                    Triple.of(p.getCaseTypeId(),
+                        p.getCaseTypeId().getJurisdiction().getId(),
+                        p.getDisplayOrder())));
+
+        accessTypeRolesDisplayOrder.keySet()
+            .forEach(triple -> {
+                if (accessTypeRolesDisplayOrder.get(triple).size() > 1) {
+                    String errorMessage = String.format("'%s' must be unique across all Case Types for "
+                            + "a given Jurisdiction in the sheet '%s'",
+                        ColumnName.DISPLAY_ORDER, SheetName.ACCESS_TYPE_ROLES);
+
+                    if (!alreadyReportedError(validationResult, errorMessage)) {
+                        validationResult.addError(new ValidationError(errorMessage) {});
+                    }
+                }
+            });
+    }
+
+    private boolean alreadyReportedError(ValidationResult validationResult, String errorMessage) {
+        return validationResult.getValidationErrors().stream()
+            .map(ValidationError::getDefaultMessage).anyMatch(errorMessage::contains);
+    }
+}
